codestyle: dotnet format IDE0001/IDE0002/IDE0004 (simplify names, member access, remove casts)
148 files. Excluded Libraries/Opc.Ua.Gds.Server.Common/Hosting/OpcUaGdsServerBuilderExtensions.cs because the simplifier inserts merge-conflict TODO markers when the simplification result differs between TFMs (same class of bug as IDE0240). IDE0005 also excluded from this batch for the same reason — handled separately.

+++ Applications/Quickstarts.Servers/ReferenceServer/ReferenceServerConfigurationNodeManager.cs
@@ -71,7 +71,7 @@ namespace Quickstarts.ReferenceServer
             await base.CreateAddressSpaceAsync(externalReferences, cancellationToken)
                 .ConfigureAwait(false);
 
-            ushort diagnosticsNamespaceIndex = (ushort)Server.NamespaceUris
+            ushort diagnosticsNamespaceIndex = Server.NamespaceUris
                 .GetIndexOrAppend(Opc.Ua.Namespaces.OpcUa + "Diagnostics");
 
             var addedNodes = new List<NodeState>();
